endpoint_exchange ver2.0
