+++ app/exchanges/routes/__init__.py
@@ -1,1 +1,2 @@
-from .exchanges import router as exchanges_router
+from .exchange import router as exchanges_router
+from .currency import router as currencies_router
